drop the kernel module
vfio-user is on track to be accepted upstream, so we shouldn't need muser.ko.
This patch removes it since keeping it in the source is a maintanance burden.
The last version is stashed in https://github.com/nutanix/muser/tree/kmod. If
there is a valid use case we can reconsider reinstating it.

Signed-off-by: Thanos Makatos <thanos.makatos@nutanix.com>